requirements.txt
requests>=2.31.0
pandas>=2.1.0
pandas_ta>=0.3.14b1
supabase>=2.0.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
schedule>=1.2.0
